% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fibanocci.R
\name{fibanocci_1}
\alias{fibanocci_1}
\title{Compute n-th fibanocci number}
\usage{
fibanocci_1(n = 1)
}
\arguments{
\item{n}{the fibanocci number you wannt to compute. By default it is set to 1.}
}
\value{
The n-th fibanocci number (numeric)
}
\description{
This function computes the n-th fibanocci number using a recursive formula.
}
\examples{
fibanocci_1()
fibanocci_1(10)

}
