% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris}
\alias{iris}
\title{Multivariate dataset with measures of the iris plant.}
\format{A data frame with columns:
\describe{
 \item{Sepal.Length}{Sepal length measure of the plant.}
 \item{Sepal.Width}{Sepal width measure of the plant.}
 \item{Petal.Length}{Petal length measure of the plant.}
 \item{Petal.Width}{Petal width measure of the plant.}
 \item{Species}{Specie of the plant}
}}
\source{
Ronald Fisher, "The use of multiple measurements in taxonomic problems", 1936.
\url{https://www.kaggle.com/arshid/iris-flower-dataset}
}
\usage{
iris
}
\description{
Multivariate dataset with measures of the iris plant.
}
\examples{
\dontrun{
 iris
}
}
\keyword{datasets}
